requirements.txt
# Core scientific computing dependencies
numpy>=1.20.0
scipy>=1.7.0
matplotlib>=3.3.0

# Quantum mechanics calculations
qutip>=4.6.0

# Molecular dynamics and structure handling
openmm>=7.6.0
mdtraj>=1.9.0
biopython>=1.78

# Data handling and I/O
pandas>=1.3.0
h5py>=3.1.0
pyyaml>=5.4.0

# Numerical optimization and integration
numba>=0.54.0
cython>=0.29.0

# Plotting and visualization
seaborn>=0.11.0
plotly>=5.0.0

# Progress bars and user interface
tqdm>=4.60.0
click>=8.0.0

# System monitoring and resource management
psutil>=5.8.0

# Scientific constants and units
astropy>=4.2.0

# Parallel computing
joblib>=1.0.0
mpi4py>=3.1.0

# Optional GPU acceleration (commented out by default)
# cupy>=9.0.0

# Development and testing dependencies
pytest>=6.0.0
pytest-cov>=2.0.0
pytest-xdist>=2.0.0
pytest-timeout>=1.4.0
black>=21.0.0
flake8>=3.8.0
mypy>=0.800

# Documentation dependencies (optional)
# sphinx>=4.0.0
# sphinx-rtd-theme>=0.5.0
# nbsphinx>=0.8.0